requirements.txt
face-recognition-models==0.3.*
gTTS==2.0.*
gTTS-token==1.1.*
Pillow==6.0.*
pygame==1.9.*
pyparsing==2.4.*
python-dateutil==2.8.*
pytz==2019.*
requests==2.22.*
soupsieve==1.9.*
SpeechRecognition==3.8.*
timeloop==1.0.*
urllib3==1.25.*
xlwt==1.3.*
